# Logging improvements