--- a/substrate/client/tracing/src/logging/fast_local_time.rs
+++ b/substrate/client/tracing/src/logging/fast_local_time.rs
@@ -20,9 +20,18 @@ use chrono::{Datelike, Timelike};
 use std::{cell::RefCell, fmt, fmt::Write, time::SystemTime};
 use tracing_subscriber::fmt::time::FormatTime;
 
-/// A structure which, when `Display`d, will print out the current local time.
+/// A structure which, when `Display`d, will print out the current time either in local timezone or
+/// in UTC.
 #[derive(Debug, Clone, Copy, Eq, PartialEq, Default)]
 pub struct FastLocalTime {
+	/// Decides whenever to use UTC timezone instead.
+	///
+	/// If `true` the output will match the following `chrono` format string:
+	///   `%Y-%m-%dT%H:%M:%S%.6fZ`
+	///
+	/// This is the same format used by [`tracing_subscriber`] by default, `with_fractional` has no
+	/// effect when this option is set to `true`.
+	pub utc: bool,
 	/// Decides whenever the fractional timestamp with be included in the output.
 	///
 	/// If `false` the output will match the following `chrono` format string:
@@ -69,6 +78,7 @@ struct CachedTimestamp {
 	timezone_offset: InlineString,
 	buffer: InlineString,
 	last_regenerated_at: u64,
+	last_utc: bool,
 	last_fractional: u32,
 }
 
@@ -78,6 +88,7 @@ thread_local! {
 
 impl FormatTime for FastLocalTime {
 	fn format_time(&self, w: &mut dyn Write) -> fmt::Result {
+		const TIMESTAMP_PARTIAL_UTC_LENGTH: usize = "0000-00-00T00:00:00".len();
 		const TIMESTAMP_PARTIAL_LENGTH: usize = "0000-00-00 00:00:00".len();
 
 		let elapsed = SystemTime::now()
@@ -89,27 +100,58 @@ impl FormatTime for FastLocalTime {
 			let mut cache = cache.borrow_mut();
 
 			// Regenerate the timestamp only at most once each second.
-			if cache.last_regenerated_at != unix_time {
-				let ts = chrono::Local::now();
-				let fractional = (ts.nanosecond() % 1_000_000_000) / 1_000_000;
-				cache.last_regenerated_at = unix_time;
-				cache.last_fractional = fractional;
-				cache.buffer.length = 0;
+			if cache.last_regenerated_at != unix_time || cache.last_utc != self.utc {
+				cache.last_utc = self.utc;
 
-				write!(
-					&mut cache.buffer,
-					"{:04}-{:02}-{:02} {:02}:{:02}:{:02}.{:03}",
-					ts.year(),
-					ts.month(),
-					ts.day(),
-					ts.hour(),
-					ts.minute(),
-					ts.second(),
-					fractional
-				)?;
+				if self.utc {
+					let ts = chrono::Utc::now();
+					let fractional = (ts.nanosecond() % 1_000_000_000) / 1_000;
+					cache.last_regenerated_at = unix_time;
+					cache.last_fractional = fractional;
+					cache.buffer.length = 0;
 
-				if cache.timezone_offset.length == 0 {
-					write!(&mut cache.timezone_offset, "{}", ts.offset())?;
+					write!(
+						&mut cache.buffer,
+						"{:04}-{:02}-{:02}T{:02}:{:02}:{:02}.{:06}Z",
+						ts.year(),
+						ts.month(),
+						ts.day(),
+						ts.hour(),
+						ts.minute(),
+						ts.second(),
+						fractional
+					)?;
+				} else {
+					let ts = chrono::Local::now();
+					let fractional = (ts.nanosecond() % 1_000_000_000) / 1_000_000;
+					cache.last_regenerated_at = unix_time;
+					cache.last_fractional = fractional;
+					cache.buffer.length = 0;
+
+					write!(
+						&mut cache.buffer,
+						"{:04}-{:02}-{:02} {:02}:{:02}:{:02}.{:03}",
+						ts.year(),
+						ts.month(),
+						ts.day(),
+						ts.hour(),
+						ts.minute(),
+						ts.second(),
+						fractional
+					)?;
+
+					if cache.timezone_offset.length == 0 {
+						write!(&mut cache.timezone_offset, "{}", ts.offset())?;
+					}
+				}
+			} else if self.utc {
+				let fractional = elapsed.subsec_micros();
+
+				// Regenerate the fractional part at most once each millisecond.
+				if cache.last_fractional != fractional {
+					cache.last_fractional = fractional;
+					cache.buffer.length = TIMESTAMP_PARTIAL_UTC_LENGTH + 1;
+					write!(&mut cache.buffer, "{:06}Z", fractional)?;
 				}
 			} else if self.with_fractional {
 				let fractional = elapsed.subsec_millis();
@@ -123,12 +165,17 @@ impl FormatTime for FastLocalTime {
 			}
 
 			let mut slice = cache.buffer.as_str();
-			if !self.with_fractional {
-				slice = &slice[..TIMESTAMP_PARTIAL_LENGTH];
-			}
 
-			w.write_str(slice)?;
-			w.write_str(cache.timezone_offset.as_str())?;
+			if self.utc {
+				w.write_str(slice)?;
+			} else {
+				if !self.with_fractional {
+					slice = &slice[..TIMESTAMP_PARTIAL_LENGTH];
+				}
+
+				w.write_str(slice)?;
+				w.write_str(cache.timezone_offset.as_str())?;
+			}
 
 			Ok(())
 		})
@@ -145,18 +192,26 @@ impl fmt::Display for FastLocalTime {
 fn test_format_fast_local_time() {
 	assert_eq!(
 		chrono::Local::now().format("%Y-%m-%d %H:%M:%S%:z").to_string().len(),
-		FastLocalTime { with_fractional: false }.to_string().len()
+		FastLocalTime { utc: false, with_fractional: false }.to_string().len()
 	);
 	assert_eq!(
 		chrono::Local::now().format("%Y-%m-%d %H:%M:%S%.3f%:z").to_string().len(),
-		FastLocalTime { with_fractional: true }.to_string().len()
+		FastLocalTime { utc: false, with_fractional: true }.to_string().len()
+	);
+	assert_eq!(
+		chrono::Utc::now().format("%Y-%m-%dT%H:%M:%S%.6fZ").to_string().len(),
+		FastLocalTime { utc: true, with_fractional: false }.to_string().len()
+	);
+	assert_eq!(
+		chrono::Utc::now().format("%Y-%m-%dT%H:%M:%S%.6fZ").to_string().len(),
+		FastLocalTime { utc: true, with_fractional: true }.to_string().len()
 	);
 
 	// A simple trick to make sure this test won't randomly fail if we so happen
 	// to land on the exact moment when we tick over to the next second.
-	let now_1 = FastLocalTime { with_fractional: false }.to_string();
+	let now_1 = FastLocalTime { utc: false, with_fractional: false }.to_string();
 	let expected = chrono::Local::now().format("%Y-%m-%d %H:%M:%S%:z").to_string();
-	let now_2 = FastLocalTime { with_fractional: false }.to_string();
+	let now_2 = FastLocalTime { utc: false, with_fractional: false }.to_string();
 
 	assert!(
 		now_1 == expected || now_2 == expected,