--- a/substrate/client/tracing/src/logging/mod.rs
+++ b/substrate/client/tracing/src/logging/mod.rs
@@ -28,7 +28,7 @@ mod fast_local_time;
 mod layers;
 mod stderr_writer;
 
-pub(crate) type DefaultLogger = stderr_writer::MakeStderrWriter;
+pub(crate) type DefaultLogger = MakeStderrWriter;
 
 pub use directives::*;
 pub use sc_tracing_proc_macro::*;
@@ -41,7 +41,7 @@ use tracing_subscriber::{
 		format, FormatEvent, FormatFields, Formatter, Layer as FmtLayer, MakeWriter,
 		SubscriberBuilder,
 	},
-	layer::{self, SubscriberExt},
+	layer::SubscriberExt,
 	registry::LookupSpan,
 	EnvFilter, FmtSubscriber, Layer, Registry,
 };
@@ -104,8 +104,8 @@ where
 	N: for<'writer> FormatFields<'writer> + 'static,
 	E: FormatEvent<Registry, N> + 'static,
 	W: MakeWriter + 'static,
-	F: layer::Layer<Formatter<N, E, W>> + Send + Sync + 'static,
-	FmtLayer<Registry, N, E, W>: layer::Layer<Registry> + Send + Sync + 'static,
+	F: Layer<Formatter<N, E, W>> + Send + Sync + 'static,
+	FmtLayer<Registry, N, E, W>: Layer<Registry> + Send + Sync + 'static,
 {
 	// Accept all valid directives and print invalid ones
 	fn parse_user_directives(mut env_filter: EnvFilter, dirs: &str) -> Result<EnvFilter> {
@@ -166,12 +166,12 @@ where
 
 	// If we're only logging `INFO` entries then we'll use a simplified logging format.
 	let detailed_output = match max_level_hint {
-		Some(level) if level <= tracing_subscriber::filter::LevelFilter::INFO => false,
+		Some(level) if level <= LevelFilter::INFO => false,
 		_ => true,
 	} || detailed_output;
 
 	let enable_color = force_colors.unwrap_or_else(|| atty::is(atty::Stream::Stderr));
-	let timer = fast_local_time::FastLocalTime { with_fractional: detailed_output };
+	let timer = FastLocalTime { with_fractional: detailed_output };
 
 	let event_format = EventFormat {
 		timer,
@@ -440,7 +440,7 @@ mod tests {
 			let test_directives = "test-target=info";
 			let _guard = init_logger(&test_directives);
 
-			log::info!(target: "test-target", "{}", EXPECTED_LOG_MESSAGE);
+			info!(target: "test-target", "{}", EXPECTED_LOG_MESSAGE);
 		}
 	}
 
@@ -475,7 +475,7 @@ mod tests {
 
 	#[crate::logging::prefix_logs_with(EXPECTED_NODE_NAME)]
 	fn prefix_in_log_lines_process() {
-		log::info!("{}", EXPECTED_LOG_MESSAGE);
+		info!("{}", EXPECTED_LOG_MESSAGE);
 	}
 
 	/// This is not an actual test, it is used by the `do_not_write_with_colors_on_tty` test.