--- a/substrate/client/tracing/src/logging/mod.rs
+++ b/substrate/client/tracing/src/logging/mod.rs
@@ -95,6 +95,7 @@ fn prepare_subscriber<N, E, F, W>(
 	directives: &str,
 	profiling_targets: Option<&str>,
 	force_colors: Option<bool>,
+	use_utc: bool,
 	detailed_output: bool,
 	builder_hook: impl Fn(
 		SubscriberBuilder<format::DefaultFields, EventFormat, EnvFilter, DefaultLogger>,
@@ -171,7 +172,7 @@ where
 	} || detailed_output;
 
 	let enable_color = force_colors.unwrap_or_else(|| atty::is(atty::Stream::Stderr));
-	let timer = FastLocalTime { with_fractional: detailed_output };
+	let timer = FastLocalTime { utc: use_utc, with_fractional: detailed_output };
 
 	let event_format = EventFormat {
 		timer,
@@ -203,6 +204,7 @@ pub struct LoggerBuilder {
 	custom_profiler: Option<Box<dyn crate::TraceHandler>>,
 	log_reloading: bool,
 	force_colors: Option<bool>,
+	use_utc: bool,
 	detailed_output: bool,
 }
 
@@ -215,6 +217,7 @@ impl LoggerBuilder {
 			custom_profiler: None,
 			log_reloading: false,
 			force_colors: None,
+			use_utc: false,
 			detailed_output: false,
 		}
 	}
@@ -261,6 +264,12 @@ impl LoggerBuilder {
 		self
 	}
 
+	/// Use UTC in log output.
+	pub fn with_utc(&mut self, utc: bool) -> &mut Self {
+		self.use_utc = utc;
+		self
+	}
+
 	/// Initialize the global logger
 	///
 	/// This sets various global logging and tracing instances and thus may only be called once.
@@ -271,6 +280,7 @@ impl LoggerBuilder {
 					&self.directives,
 					Some(&profiling_targets),
 					self.force_colors,
+					self.use_utc,
 					self.detailed_output,
 					|builder| enable_log_reloading!(builder),
 				)?;
@@ -289,6 +299,7 @@ impl LoggerBuilder {
 					&self.directives,
 					Some(&profiling_targets),
 					self.force_colors,
+					self.use_utc,
 					self.detailed_output,
 					|builder| builder,
 				)?;
@@ -308,6 +319,7 @@ impl LoggerBuilder {
 				&self.directives,
 				None,
 				self.force_colors,
+				self.use_utc,
 				self.detailed_output,
 				|builder| enable_log_reloading!(builder),
 			)?;
@@ -320,6 +332,7 @@ impl LoggerBuilder {
 				&self.directives,
 				None,
 				self.force_colors,
+				self.use_utc,
 				self.detailed_output,
 				|builder| builder,
 			)?;