This PR is an improved version of https://github.com/paritytech/substrate/pull/14760

Here I changed default timestamp format in logging to include timezone offset (as requested by @koute) and added CLI option to print timestamps in UTC using the same format as the default in `tracing-subscriber`.

Before:
```
2023-09-12 02:57:03 Substrate Node    
```
After:
```
2023-09-12 02:57:03+01:00 Substrate Node    
```
UTC:
```
2023-09-12T01:57:03.061965Z Substrate Node    
```

I also did benchmarks and found that these changes introduce ~10% performance hit for local time, while UTC performance is about the same as local time was before these changes.

Before:
```
fast_local_time         time:   [18.366 ns 18.392 ns 18.421 ns]
chrono_local            time:   [320.59 ns 320.83 ns 321.07 ns]
```
After:
```
fast_local_time/local   time:   [19.919 ns 19.927 ns 19.937 ns]
fast_local_time/utc     time:   [18.582 ns 18.590 ns 18.599 ns]
```

There are also some changes to colors:
* Support for [`NO_COLOR`](https://no-color.org/), though better approach might be to remove --disable-log-color altogether and use supports-color that automatically takes this environment variable into account (see https://github.com/tokio-rs/tracing/issues/2388 and https://github.com/tokio-rs/tracing/issues/2214).
* On Windows colors are disabled by default because otherwise they make logs hard to read since PowerShell doesn't handle escape characters teh way *nix shells do

I recommend reviewing individual commits rather than final diff for convenience.