@nazar-pc 2023-11-21
@koute I'd really like to find the design which works for everyone and upstream it. Do you have any fresh thoughts on how to approach this?