@nazar-pc 2024-01-10
Closing as I no longer have interest in driving this forward, our chain is switching away from Substrate's runner and as such we can use normal `tracing-subscriber` directly without fighting Substrate's logging system.